final touch

+++ src/app.controller.ts
@@ -96,16 +96,20 @@ export class AppController {
 
   @Get('/cleanall')
   async cleanSpams() {
-    getRepository(MatchingSpams)
-      .createQueryBuilder()
-      .update()
-      .set({ accept_list: [], decline_list: [], remove_list: [], matched_list: [] })
-      .execute();
+    if (process.env.NODE_ENV === 'development') {
+      getRepository(MatchingSpams)
+        .createQueryBuilder()
+        .update()
+        .set({ accept_list: [], decline_list: [], remove_list: [], matched_list: [] })
+        .execute();
 
-    getRepository(MatchedDuos).createQueryBuilder().delete().where('id > 0').execute();
-    getRepository(MatchedDuosNotifications).createQueryBuilder().delete().where('id > 0').execute();
-    getRepository(MatchingLobby).createQueryBuilder().delete().where('id > 0').execute();
-    return 'cleaned';
+      getRepository(MatchedDuos).createQueryBuilder().delete().where('id > 0').execute();
+      getRepository(MatchedDuosNotifications).createQueryBuilder().delete().where('id > 0').execute();
+      getRepository(MatchingLobby).createQueryBuilder().delete().where('id > 0').execute();
+      return 'cleaned';
+    }
+
+    return 'not so fast';
   }
 
   @Get('/relations')
